REFACT: 질문, 답변, 댓글 모델 각각 분류
- Question, Answer, Comment 모델을 각각의 파일에서 생성
- __init__.py 파일에 모델들 포함
