+++ pybo/models/__init__.py
@@ -1,0 +1,5 @@
+from .question import Question
+from .answer import Answer
+from .comment import Comment
+
+__all__ = ['Question', 'Answer', 'Comment']
