+++ pybo/models/answer.py
@@ -1,0 +1,13 @@
+from django.contrib.auth.models import User
+from django.db import models
+
+from .question import Question
+
+
+class Answer(models.Model):
+    author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='author_answer')
+    question = models.ForeignKey(Question, on_delete=models.CASCADE)
+    content = models.TextField()
+    create_date = models.DateTimeField()
+    modify_date = models.DateTimeField(null=True, blank=True)
+    voter = models.ManyToManyField(User, related_name='voter_answer')
