+++ pybo/models/comment.py
@@ -1,0 +1,15 @@
+from django.contrib.auth.models import User
+from django.db import models
+from django.utils import timezone
+
+from .question import Question
+from .answer import Answer
+
+
+class Comment(models.Model):
+    author = models.ForeignKey(User, on_delete=models.CASCADE)
+    content = models.TextField()
+    create_date = models.DateTimeField()
+    modify_date = models.DateTimeField(null=True, blank=True)
+    question = models.ForeignKey(Question, null=True, blank=True, on_delete=models.CASCADE)
+    answer = models.ForeignKey(Answer, null=True, blank=True, on_delete=models.CASCADE)
