+++ pybo/models/question.py
@@ -33,21 +33,3 @@ class Question(models.Model):
 
     def delete_question(self) -> None:
         self.delete()
-
-
-class Answer(models.Model):
-    author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='author_answer')
-    question = models.ForeignKey(Question, on_delete=models.CASCADE)
-    content = models.TextField()
-    create_date = models.DateTimeField()
-    modify_date = models.DateTimeField(null=True, blank=True)
-    voter = models.ManyToManyField(User, related_name='voter_answer')
-
-
-class Comment(models.Model):
-    author = models.ForeignKey(User, on_delete=models.CASCADE)
-    content = models.TextField()
-    create_date = models.DateTimeField()
-    modify_date = models.DateTimeField(null=True, blank=True)
-    question = models.ForeignKey(Question, null=True, blank=True, on_delete=models.CASCADE)
-    answer = models.ForeignKey(Answer, null=True, blank=True, on_delete=models.CASCADE)
